Added MutationObserver
A MutationObserver was added in the situation that someone may have a node created, but not yet created to the DOM. When they add it to the DOM, the image will not load because a scroll hasn't happened. The window size "may" have changed, but that's not always the case. In comes the MutationObserver. We can use this to observe the DOM of a certain element, in this case the body, and then detect if nodes were added or removed, and then perform a lazy load image check.